Fix ChartSatistic, rendering
